feat(groups): display created and joined groups depending on user role and present the possibily of adding or joining a group
