+++ images/app/package.json
@@ -13,6 +13,7 @@
     "@expo/metro-runtime": "~3.2.1",
     "@expo/next-adapter": "^6.0.0",
     "@react-native-async-storage/async-storage": "^1.23.1",
+    "@react-native-picker/picker": "^2.7.6",
     "@react-navigation/native": "^6.1.17",
     "@react-navigation/stack": "^6.3.29",
     "expo": "^51.0.2",
@@ -25,6 +26,7 @@
     "expo-splash-screen": "^0.27.4",
     "expo-status-bar": "~1.12.1",
     "firebase": "^10.10.0",
+    "jwt-decode": "^4.0.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "^0.74.1",
@@ -32,6 +34,8 @@
     "react-native-linear-gradient": "^2.8.3",
     "react-native-safe-area-context": "^4.10.1",
     "react-native-screens": "^3.31.1",
+    "react-native-svg": "^15.3.0",
+    "react-native-svg-uri": "^1.2.3",
     "react-native-web": "~0.19.6"
   },
   "devDependencies": {
